backend/requirements.txt
Django>=4.2,<5.1
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
psycopg2-binary>=2.9,<3.0
beautifulsoup4>=4.12,<5.0
pypdf>=4.2,<5.0
